httpd/mi_http: allign callback functions with the latest microhttpd API
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@9311 689a6050-402a-0410-94f2-e92a70836424

+++ modules/httpd/httpd_load.h
@@ -91,7 +91,7 @@ typedef void (httpd_acces_handler_cb) (void *cls, void *connection, const char *
  *  (possibly allowing the client to perform additional
  *  requests using the same TCP connection).
  */
-typedef int (httpd_flush_data_cb) (void *cls, uint64_t pos, char *buf, int max);
+typedef ssize_t (httpd_flush_data_cb) (void *cls, uint64_t pos, char *buf, size_t max);
 
 /**
  * Callback to be run in order to initialize process specific data
